pyproject.toml
[tool.poetry]
name = "vk-botfarm-service"
version = "0.1.0"
description = "Тестовое задание на стажировку в VK - сервис ботофермы VK"
authors = ["Дмитрий Панин <dimitri-05@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.3"}

[tool.poetry.group.database.dependencies]
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.1"

[tool.poetry.group.security.dependencies]
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
black = "^24.4.2"

[tool.poetry.group.test.dependencies]
httpx = "^0.27.0"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"


[tool.isort]
profile = "black"
line_length = 100
extend_skip = ["migrations"]
skip_gitignore = true


[tool.black]
line-length = 100
extend-exclude = "migrations"


[tool.pytest.ini_options]
asyncio_mode = "auto"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
